Approved free license for the Qt package
This is done so that the entire rootfs image build process is not interrupted
part way by the Qt package asking the user which license to select (open-source
vs. commercial). The open source license will be automatically selected.

Did a `make menuconfig', went to Package Selection for the target -> Graphic
libraries and applications (graphic/text) -> Qt, and selected "Approve free
license".

This turns on symbol BR2_PACKAGE_QT_LICENSE_APPROVED.