Turned on GDB and gdbserver in menuconfig
Buildroot now installs the GNU Debugger (GDB) on the development machine, and
gdbserver (the GDB server) on the target (Smart Instrument board).

To debug the executable, first do a `make clean', then a `make debug' from the
$SMARTINSTR_INSTRFW directory. (Building debug versions of the Shared Object
Libraries (FIUsbInt.dll.so etc.) is not yet supported by the Makefile.)

To launch the debug session, follow the instructions in the Buildroot manual,
section 8.11.2, "Using gdb in Buildroot" (manual version 2015.05). (Replace
`foo' with `instrfw-debug'.)

Note: I was not able to locate the following file:
   <buildroot>/output/staging/usr/share/buildroot/gdbinit
   (probably due to the fact that we are using an older version of Buildroot)

I could launch gdbserver and gdb host, and establish communication between the
two, but I was not able to successfully run a debug session.